Restore main theme (#380)
* readded support for background image

* testing waves theme

* Update themes.xml

* updated themes

* updated themes

* updated themes

* fixed plugin compatiblity check

apparently 2.6 float < 2.6 double

* updated themes

* updated themes

* updated theme picker

* updated waves theme

* Delete theme_colors_circle.xml

* formatted code

+++ app/app/src/main/java/com/github/livingwithhippos/unchained/data/model/PluginRepository.kt
@@ -57,6 +57,6 @@ data class PluginVersion(
     @ColumnInfo(name = "plugin_repository") val repository: String,
     @ColumnInfo(name = "plugin") val plugin: String,
     @ColumnInfo(name = "version") val version: Float,
-    @ColumnInfo(name = "engine") val engine: Double,
+    @ColumnInfo(name = "engine") val engine: Float,
     @ColumnInfo(name = "plugin_link") val link: String
 )
